+++ doc/CHANGES.rst
@@ -28,6 +28,7 @@ Futur
 * [REF]: adapter of mapper_args and table_args become an ``anyblok.model.plugin``
 * [REF]: **event** become an ``anyblok.model.plugin``
 * [REF]: **sqlachemy event** become an ``anyblok.model.plugin``
+* [REF]: **cache** and **classmethod_cache** become an ``anyblok.model.plugin``
 
 0.9.10 (2017-09-23)
 -------------------